+++ src/dftbp/dftbplus/mainio.F90
@@ -85,7 +85,8 @@ module dftbp_dftbplus_mainio
   public :: openOutputFile
   public :: writeDetailedOut1, writeDetailedOut2, writeDetailedOut2Dets, writeDetailedOut3
   public :: writeDetailedOut4, writeDetailedOut5, writeDetailedOut6, writeDetailedOut7
-  public :: writeDetailedOut8, writeDetailedOut9, writeDetailedOut10, permitivityPrint
+  public :: writeDetailedOut8, writeDetailedOut9, writeDetailedOut10
+  public :: permitivityPrint
   public :: writeMdOut1, writeMdOut2
   public :: writeCharges
   public :: writeEsp